requirements.txt
Flask==3.0.0
pypdf==3.17.4
python-docx==1.1.0
reportlab==4.2.2
plotly==5.18.0
textblob==0.18.0
pandas==2.2.0
numpy==1.26.4
Pillow==10.2.0
nltk==3.8.1
wordcloud==1.9.3
groq==0.4.2
python-dotenv==1.0.1
requests==2.32.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
gunicorn==21.2.0